Moved Register.js component and fixed minor bug
